requirements.txt
numpy==1.22.4
pandas==1.4.2
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
six==1.16.0
